requirement.txt
Django==3.2.12
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
